refactor: verify if account exists before delete and throw error if not

+++ src/main/java/com/raponi/blog/application/service/account/DeleteAccountService.java
@@ -1,9 +1,13 @@
 package com.raponi.blog.application.service.account;
 
+import java.util.Optional;
+
 import org.springframework.stereotype.Service;
 
+import com.raponi.blog.domain.model.Account;
 import com.raponi.blog.domain.usecase.account.DeleteAccountUseCase;
 import com.raponi.blog.infrastructure.persistence.repository.AccountRepository;
+import com.raponi.blog.presentation.errors.AccountNotFound;
 
 @Service
 public class DeleteAccountService implements DeleteAccountUseCase {
@@ -16,8 +20,12 @@ public class DeleteAccountService implements DeleteAccountUseCase {
 
   @Override
   public String handle(String accountId) {
-    this.accountRepository.deleteById(accountId);
-    return "Conta deletada com sucesso";
+    Optional<Account> account = this.accountRepository.findById(accountId);
+    if (account.isPresent()) {
+      this.accountRepository.deleteById(accountId);
+      return "Conta deletada com sucesso";
+    }
+    throw new AccountNotFound("id equals " + accountId);
   }
 
 }
